# server/api/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Visora AI P1", "version": "0.2.0", "description": "本地开发账号服务与运营后台。写请求必须通过浏览器 Origin 精确白名单验证，缺失或不匹配返回403 FORBIDDEN。写方法包含 POST/PATCH/PUT/DELETE，四者都要求 Origin、X-Visora-Request 与 Content-Type 三道校验。浏览器只调用本 origin 的 /api；开发用户端/Admin统一使用127.0.0.1，不混用localhost。代理保留 Origin、Cookie 和 Set-Cookie，不让浏览器直接请求API_TARGET。项目云存储、平台生成任务与积分消耗结算仍未接入；积分流水中的 consume 类型由后续生成迭代写入，当前没有写入方。" },
  "security": [{ "session": [] }],
  "paths": {
...
    "/api/admin/teams": {
      "get": {
        "operationId": "listTeams", "description": "仅管理员；团队列表，createdAt/_id 倒序。allocatedThisMonth 由积分流水聚合得到（UTC 自然月）。",
        "parameters": [{ "$ref": "#/components/parameters/Cursor" }],
        "responses": { "200": { "$ref": "#/components/responses/TeamPage" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      },
      "post": {
        "operationId": "createTeam", "description": "仅管理员；创建团队并把指定用户设为团队管理员成员。团队名唯一，重复返回 409 NAME_EXISTS。新团队积分池为 0。",
        "parameters": [{ "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/TeamCreate" },
        "responses": { "201": { "$ref": "#/components/responses/Team" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/teams/{id}/dissolve": {
      "post": {
        "operationId": "dissolveTeam", "description": "仅管理员；解散团队。按需求对照 C07 只把状态改为 dissolved 并冻结后续积分操作，积分池余额与成员关系保留，不清空积分、不删除素材。破坏性清算待 D09 确认后再实现。",
        "parameters": [{ "$ref": "#/components/parameters/PathID" }, { "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false } } } },
        "responses": { "200": { "$ref": "#/components/responses/Team" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/teams/{id}/members": {
      "get": {
        "operationId": "listTeamMembers", "description": "仅管理员；团队成员与各自已分配积分。",
        "parameters": [{ "$ref": "#/components/parameters/PathID" }],
        "responses": { "200": { "$ref": "#/components/responses/TeamMemberPage" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      },
      "post": {
        "operationId": "addTeamMember", "description": "仅管理员；加入成员，重复加入返回 409 MEMBER_EXISTS。团队已解散返回 409 TEAM_INACTIVE。",
        "parameters": [{ "$ref": "#/components/parameters/PathID" }, { "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["userId"], "additionalProperties": false, "properties": { "userId": { "type": "string" } } } } } },
        "responses": { "201": { "$ref": "#/components/responses/TeamMember" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/teams/{id}/members/{userId}": {
      "delete": {
        "operationId": "removeTeamMember", "description": "仅管理员；移出成员。团队管理员不能移出；成员仍有已分配积分时返回 409 MEMBER_HAS_POINTS，要求先撤销，避免积分悬空。DELETE 同样需要 Origin 与请求头校验。",
        "parameters": [{ "$ref": "#/components/parameters/PathID" }, { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }, { "$ref": "#/components/parameters/RequestHeader" }],
        "responses": { "200": { "$ref": "#/components/responses/TeamMember" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/teams/{id}/points/topup": {
      "post": {
        "operationId": "topUpTeam", "description": "仅管理员；平台向团队积分池充入积分，写入 scope=team_pool 的流水。",
        "parameters": [{ "$ref": "#/components/parameters/PathID" }, { "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/TeamTopUp" },
        "responses": { "200": { "$ref": "#/components/responses/Team" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/teams/{id}/points/allocate": {
      "post": {
        "operationId": "allocateTeamPoints", "description": "仅管理员；从团队池分配到成员，同时减少池余额并增加成员已分配积分。池余额不足返回 409 INSUFFICIENT_POINTS。",
        "parameters": [{ "$ref": "#/components/parameters/PathID" }, { "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/TeamAllocation" },
        "responses": { "200": { "$ref": "#/components/responses/TeamMember" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/teams/{id}/points/revoke": {
      "post": {
        "operationId": "revokeTeamAllocation", "description": "仅管理员；撤销成员已分配积分并退回团队池。成员已分配不足返回 409 INSUFFICIENT_POINTS。",
        "parameters": [{ "$ref": "#/components/parameters/PathID" }, { "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/TeamAllocation" },
        "responses": { "200": { "$ref": "#/components/responses/TeamMember" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
...
  },
  "components": {
    "headers": { "NoStore": { "schema": { "type": "string", "const": "no-store" } } },
    "securitySchemes": { "session": { "type": "apiKey", "in": "cookie", "name": "visora_session" } },
    "parameters": {
      "RequestHeader": { "name": "X-Visora-Request", "in": "header", "required": true, "schema": { "type": "string", "const": "1" } },
      "Cursor": { "name": "cursor", "in": "query", "schema": { "type": "string" } },
      "PathID": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
    },
    "requestBodies": {
...
      "TeamCreate": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name", "adminUserId"], "additionalProperties": false, "properties": { "name": { "type": "string", "maxLength": 60 }, "adminUserId": { "type": "string" } } } } } },
      "TeamTopUp": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["amount"], "additionalProperties": false, "properties": { "amount": { "type": "integer", "minimum": 1, "maximum": 100000000 }, "note": { "type": "string", "maxLength": 200 } } } } } },
      "TeamAllocation": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["userId", "amount"], "additionalProperties": false, "properties": { "userId": { "type": "string" }, "amount": { "type": "integer", "minimum": 1, "maximum": 100000000 }, "note": { "type": "string", "maxLength": 200 } } } } } },
...
    },
    "schemas": {
...
      "Team": { "type": "object", "required": ["id", "name", "adminUserId", "adminEmail", "memberCount", "pointsPool", "allocatedTotal", "allocatedThisMonth", "status", "createdAt"], "additionalProperties": false, "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "adminUserId": { "type": "string" }, "adminEmail": { "type": "string" }, "memberCount": { "type": "integer" }, "pointsPool": { "type": "integer", "description": "团队积分池余额。" }, "allocatedTotal": { "type": "integer", "description": "累计已分配到成员的积分。" }, "allocatedThisMonth": { "type": "integer", "description": "UTC 自然月内已分配积分。" }, "status": { "type": "string", "enum": ["active", "dissolved"] }, "createdAt": { "type": "string", "format": "date-time" } } },
      "TeamMember": { "type": "object", "required": ["userId", "email", "role", "allocatedPoints", "createdAt"], "additionalProperties": false, "properties": { "userId": { "type": "string" }, "email": { "type": "string" }, "role": { "type": "string", "enum": ["admin", "member"] }, "allocatedPoints": { "type": "integer" }, "createdAt": { "type": "string", "format": "date-time" } } },
...
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "object", "required": ["code", "message"], "properties": { "code": { "type": "string", "description": "400 INVALID_INPUT；401 INVALID_CREDENTIALS（登录）/UNAUTHENTICATED（会话）；403 FORBIDDEN（角色/Origin/header/Content-Type）；404 NOT_FOUND/MEMBER_NOT_FOUND/PLAN_NOT_FOUND；405 METHOD_NOT_ALLOWED；409 EMAIL_EXISTS/NAME_EXISTS/USER_PROTECTED/TEAM_INACTIVE/MEMBER_EXISTS/MEMBER_HAS_POINTS/INSUFFICIENT_POINTS/ALREADY_DECIDED/SELF_ACTION/LAST_ADMIN；413 PAYLOAD_TOO_LARGE；503 SERVICE_UNAVAILABLE。", "examples": ["INVALID_INPUT", "INVALID_CREDENTIALS", "UNAUTHENTICATED", "FORBIDDEN", "NOT_FOUND", "MEMBER_NOT_FOUND", "PLAN_NOT_FOUND", "METHOD_NOT_ALLOWED", "EMAIL_EXISTS", "NAME_EXISTS", "USER_PROTECTED", "TEAM_INACTIVE", "MEMBER_EXISTS", "MEMBER_HAS_POINTS", "INSUFFICIENT_POINTS", "ALREADY_DECIDED", "SELF_ACTION", "LAST_ADMIN", "PAYLOAD_TOO_LARGE", "SERVICE_UNAVAILABLE"] }, "message": { "type": "string" } } } } }
    },
    "responses": {
...
      "Team": { "description": "团队", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "type": "object", "required": ["team"], "properties": { "team": { "$ref": "#/components/schemas/Team" } } } } } },
      "TeamMember": { "description": "团队成员", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "type": "object", "required": ["member"], "properties": { "member": { "$ref": "#/components/schemas/TeamMember" } } } } } },
...
      "TeamPage": { "description": "团队分页", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "type": "object", "required": ["items", "nextCursor"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Team" } }, "nextCursor": { "type": ["string", "null"] } } } } } },
      "TeamMemberPage": { "description": "团队成员列表", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/TeamMember" } } } } } } },
...
      "Error": { "description": "受控错误；不回显数据库错误或凭据", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    }
  }
}